requirements.txt
asgiref==3.7.2
biopython==1.81
Django==4.2.5
elementpath==4.1.5
numpy==1.26.0
sqlparse==0.4.4
